% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoGPS.R
\docType{package}
\name{CoGPS-package}
\alias{CoGPS-package}
\alias{CoGPS}
\title{A novel method for deciphering context-specific gene programs using single-cell and spatial transcriptomics data}
\description{
CoGPS provides a computational framework for characterizing context-specific gene programs within scRNA-seq and ST datasets.
}
