% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeCEP.R
\docType{package}
\name{DeCEP-package}
\alias{DeCEP-package}
\alias{DeCEP}
\title{A novel method for deciphering context-specific gene programs using single-cell and spatial transcriptomics data}
\description{
DeCEP provides a computational framework for characterizing context-specific gene programs using scRNA-seq and ST data.
}
